// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/ConsumoRollosCorrugadora/IParametros.ts
export interface IParametros {

    FechaIni:string;
    FechaFin:string;
    chkTodosTurnos: number;
    Turno: string;
    CodigoRollo: string;
    Clase : string;
    SubClase : string;
    Ancho: number;
    PesoInicial: number;
    PesoInicio: number;
    PesoFinal: number;
    RadioFinal: number;
    ChkOps: number;
    OpcionTipoRolloSeleccionado: string;
    TipoOrdenacion: string;
      
}